requirements.text.txt
pytest==8.3.2
pytest-asyncio==0.24.0
pytest-httpx==0.30.0
time-machine==2.15.0
